# internal/dto/login-dto.go
package dto

// LoginDTO is a model that user by client when POST from '/login'
type LoginDTO struct {
	Email    string `json:"email" form:"email" binding:"required,email"`
	Password string `json:"password" form:"password" binding:"required,min=6"`
}

type DataAlex struct {
	ID    uint64 `json:"id"`
	Token string `json:"string"`
}