# internal/dto/register-dto.go
package dto

type RegisterDTO struct {
	Name       string `json:"name" form:"name" binding:"required"`
	Email      string `json:"email" form:"email" binding:"required,email"`
	Password   string `json:"password" form:"password" binding:"required,min=6"`
	CodeVerify int    `json:"code_verify"`
}